move withSIGTERMCancel to cobra command
