+++ pkg/cmd/daemon.go
@@ -1,6 +1,10 @@
 package cmd
 
 import (
+	"os"
+	"os/signal"
+	"syscall"
+
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
@@ -34,13 +38,34 @@ Currently, stripe daemon only supports a subset of CLI commands. Documentation i
 	return dc
 }
 
+func withSIGTERMCancel(ctx context.Context, onCancel func()) context.Context {
+	// Create a context that will be canceled when Ctrl+C is pressed
+	ctx, cancel := context.WithCancel(ctx)
+
+	interruptCh := make(chan os.Signal, 1)
+	signal.Notify(interruptCh, os.Interrupt, syscall.SIGTERM)
+
+	go func() {
+		<-interruptCh
+		onCancel()
+		cancel()
+	}()
+	return ctx
+}
+
 func (dc *daemonCmd) runDaemonCmd(cmd *cobra.Command, args []string) error {
 	srv := rpcserver.New(&rpcserver.Config{
 		Port: dc.port,
 		Log:  log.StandardLogger(),
 	})
 
-	err := srv.Run(context.Background())
+	ctx := withSIGTERMCancel(context.Background(), func() {
+		log.WithFields(log.Fields{
+			"prefix": "cmd.daemonCmd.runDaemonCmd",
+		}).Debug("Ctrl+C received, cleaning up...")
+	})
+
+	err := srv.Run(ctx)
 	if err != nil {
 		return err
 	}
